Meta model for working with content in rdflib
The metamodel is a starting point to make it easier to work with data
stored in rdflib.  The idea is that it would become sufficiently
simple to map contents in the triplestore to contents in native models
in such a way that we can use them as if they were a native part of
the Ember ecosystem.

This metamodel is a PoC.  There is a still a long way to go to get
this to production.