Support for single-direction hasMany relationship
The hasMany relationship already allows us to walk the data in one
direction.  This way of processing the data removes the bulk of manual
fetching of content.  The factory method that existed earlier on now
finds the relevant PropertyGroup instances, instantiates them, and
sorts them.

We are not fully ready to connect this into a model for the Form
itself which would dynamically load based on the contents of the
form.  Further experimentation will show how far this sort of model
can be used for interaction when combined with specific form
contents.