Integration with Ember Inspector
This was a bit tricky to execute at first.  Pointers by @lifeart were
quintessential for the implementation.

Some callbacks have been added to the implementation to help with the
implementation of this service.  The current implementation will
overlap with Ember Data's implementation and some kind of combining
data-adapter would need to be built to match these two worlds.

The implementation is crude.  The goal is not to end with a polished
product at this stage.  The code will be cleaned up when it's pulled
into a separate library.