Support for inverse relationships and type tracking
When a new instance is created, a triple is added to the triplestore
which indicates the type of the instance.  Making this work required
the creation of annotations rather than setting the variables
directly.

The system now also understands bidirectional relationships.  They are
represented in triples as you'd expect, but the relationship can be
fetched an reasoned on based on instances.