Depend on ember-rdflib and autosave option
ember-rdflib does the dirtywork of bundling rdflib.  It makes much
more sense to use that rather than to use our own local version.

It is now possible to indicate which triples should be handled through
Pusher (thereby letting you update a SOLID pod) and which triples can
be updated locally in rdflib.  This approach allows for transparent
updating of the store.

There are other strategies which are likely to be desired.  Possibly
using multiple triplestores.  In any case, the current construction
lets us experiment with applications and that's our first goal with
this mini setup application.