Decide target graphs based on SOLID location
The location for a certain type of data can be supplied in a SOLID
pod.  This allows the user to decide where certain data should land.
Our implementation now reads such information from the solid pod.

There are a bunch of rough edges in this implementation but the
general idea exists.