Move data to where it belongs and more robust fetching
The fetching made things crash, it also fetched inverse relations from
the wrong end in some cases.  Both of these have been resolved.

We implemented a quick fetching component to help us abstract dirty
code during initial development.  This has now been moved into the
route.  We should clean up the code in the route to ensure we can more
cleanly fetch this sort of data and populate the store with
information on the world we already know.