Introducing the ForkingStore
The ForkingStore knows which triples have been inserted and which
have been removed.  As such, it can temporarily record changes which
are then later written to the backend.

We're still in poc-mode here.  Most of the things are proof of
concept.  That being said, it should store the right data when
requested and bugs should be discovered quite easily due to the
limited supported API surface.